+++ parea/evals/chat/goal_success_ratio.py
@@ -1,6 +1,5 @@
-from typing import Callable, Optional
-
 import json
+from typing import Optional, Callable
 
 from parea.evals.utils import call_openai
 from parea.schemas.log import Log
